Use MultiWriterIDGenerator for account data
